cli: drop the apply subcommand
This removes `torcx apply` subcommand, which has now been fully
replaced by the generator.
